cleaner overrides, better error handling

+++ win64/tools/toggle-proxy.vbs
@@ -13,7 +13,8 @@ Const sProxyServer   = "HKCU\Software\Microsoft\Windows\CurrentVersion\Internet 
 Const sLocal = "<local>"
 Const sSocks="socks=127.0.0.1:12345"
 Dim aOverrides, sOverrides 
-aOverrides = array("corp.viverae.com", "helpdesk", "172.20.119.210", "dev.viverae.com", "viveraeconnect.com", "184.73.235.12", "107.22.210.163", "184.73.255.87", "204.236.224.158", "23.21.122.20", "23.23.126.165", "54.243.248.138", "23.21.62.199", sLocal)
+'aOverrides = array("corp.viverae.com", "helpdesk", "172.20.119.210", "dev.viverae.com", "viveraeconnect.com", "184.73.235.12", "107.22.210.163", "184.73.255.87", "204.236.224.158", "23.21.122.20", "23.23.126.165", "54.243.248.138", "23.21.62.199", sLocal)
+aOverrides = array("*.corp.viverae.com", "corp.viverae.com", "helpdesk", "172.20.119.210", "viveraeconnect.com", sLocal)
 sOverrides = Join(aOverrides, ";")
 
 'Display current setting
@@ -87,17 +88,21 @@ End Sub
 Sub KillByShortName(objWMIService, cmdShort)
   Dim colProcess, objProcess, intrc, pid
   Set colProcess = objWMIService.ExecQuery("Select * from Win32_Process where Name = '" & cmdShort & "'")
-  For Each objProcess in colProcess
-    WScript.Sleep 1500
-    If IsObject(objProcess) Then
-      pid = objProcess.processid
-      intrc = objProcess.Terminate()
-      If intrc = 0 Then
-        WScript.Echo "... Killed " & pid
-      Else
-        WScript.Echo "... Could not kill " & intrc & " Reason code: " & intrc
-      End If
-    End If
-  Next
+  If colProcess.Count > 0 Then
+    On Error Resume next
+      For Each objProcess in colProcess
+        WScript.Sleep 1500
+        If IsObject(objProcess) Then
+          pid = objProcess.processid
+          intrc = objProcess.Terminate()
+          If intrc = 0 Then
+            WScript.Echo "... Killed " & pid
+          Else
+            WScript.Echo "... Could not kill " & intrc & " Reason code: " & intrc
+          End If
+        End If
+      Next
+    On Error goTo 0
+  End If
 End Sub
